Split Equipment list and EquipType
